pyproject.toml
[tool.poetry]
name = "repl-nix-webhookdatastorage"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask-sqlalchemy = "^3.1.1"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
flask-login = "^0.6.3"
flask-socketio = "5.3.7"
gunicorn = "23.0.0"
eventlet = "0.37.0"
pytz = "2024.2"
requests = "2.32.3"
flask-caching = "2.3.0"
nltk = "3.9.1"
scikit-learn = "1.5.2"
flask-oauthlib = "0.9.6"
google-auth-oauthlib = "1.2.1"
google-auth = "2.35.0"
authlib = "1.3.2"
langchain = "0.3.1"
langchain-openai = "0.2.1"
langchain-anthropic = "0.2.1"
langchain-core = "0.3.6"
pydantic = "2.9.2"
apscheduler = "3.10.4"
openai = "1.50.2"
emoji = "2.13.2"
flask-apscheduler = "1.13.1"
sqlalchemy = "2.0.35"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
